Struct spinoso_time::tzrs::Offset

pub struct Offset { /* private fields */ }

Represents the number of seconds offset from UTC.

Implementations

impl Offset

pub fn utc() -> Self

Generate a UTC based offset.

Examples

let offset = Offset::utc();
assert!(offset.is_utc());

let time = Time::new(2022, 7, 29, 12, 36, 0, 0, offset)?;
assert!(time.is_utc());

pub fn local() -> Self

Generate an offset based on the detected local time zone of the system.

Detection is done by tzdb::local_tz, and if it fails will return a GMT timezone.

The system timezone is detected on the first call to this function and will be constant for the life of the program.

Examples

let offset = Offset::local();
assert!(!offset.is_utc());

let time = Time::new(2022, 7, 29, 12, 36, 0, 0, offset)?;
assert!(!time.is_utc());

pub fn fixed(offset: i32) -> Result<Self, TimeError>

Generate an offset with a number of seconds from UTC.

Examples

let offset = Offset::fixed(6600)?; // +0150
assert!(!offset.is_utc());

let time = Time::new(2022, 7, 29, 12, 36, 0, 0, offset)?;
assert!(!time.is_utc());

The offset must be in range:

let offset = Offset::fixed(500_000); // +0150
assert!(offset.is_err());

Errors

Return a TimeError::TzOutOfRangeError when outside of range of acceptable offset of MIN_OFFSET_SECONDS to MAX_OFFSET_SECONDS.

pub fn is_utc(&self) -> bool

Returns whether this offset is UTC.

Examples

let offset = Offset::utc();
assert!(offset.is_utc());

let offset = Offset::fixed(6600)?; // +0150
assert!(!offset.is_utc());