Crate spinoso_time

Time is an abstraction of dates and times.

This module implements the Time class from Ruby Core.

In Artichoke, Time is represented as a 64-bit signed integer of seconds since January 1, 1970 UTC (the Unix Epoch) and an unsigned 32-bit integer of subsecond nanoseconds. This allows representing roughly 584 billion years.

You can use this class in your application by accessing it directly. As a Core class, it is globally available:

Time.now

This implementation of Time is dependant on the selected feature. The chrono feature uses the chrono crate, and the tzrs feature uses the tzdb for getting the local timezone information, and combines with the tz-rs crate to generate the time.

Crate features

This crate supports two backends which are mutually exclusive to each other. These backends can be selected using the following features:

chrono which is backed by the chrono crate

tzrs which is backed by the tz-rs crate

Additional features

tzrs-local (enabled by default, implicitly enables tzrs)

This enables the detection of the system timezone. If disabled, defaults to GMT (not UTC)

This crate requires std, the Rust Standard Library.

Modules

chrono

A Time struct backed by the chrono rust crate

tzrs

A Time struct backed by the tz-rs rust crate

Constants

MICROS_IN_NANO

Number of microseconds in one nanosecond.

NANOS_IN_SECOND

Number of nanoseconds in one second.