Crate spinoso_time

Time is an abstraction of dates and times.

This module implements the Time class from Ruby Core.

In Artichoke, Time is represented as a 64-bit signed integer of seconds since January 1, 1970 UTC (the Unix Epoch) and an unsigned 32-bit integer of subsecond nanoseconds. This allows representing roughly 584 billion years.

You can use this class in your application by accessing it directly. As a Core class, it is globally available:

Time.now

This implementation of Time supports the system clock via the chrono crate.

Crate features

This crate requires std, the Rust Standard Library.

Structs

Time
Implementation of Ruby Time, a timezone-aware datetime, based on chrono.

ToA
Serialized representation of a timestamp using a ten-element array of datetime components.

Enums

ComponentOutOfRangeError
Error returned when constructing a Time from a ToA.

Offset
Timestamp offsets from UTC.