spinoso_time

Module tzrs

A Time struct backed by the tz-rs crate.

Structs

IntOverflowError

Error that indicates a given operation has resulted in an integer overflow.

Offset

Represents the number of seconds offset from UTC.

Time

Implementation of Ruby Time, a timezone-aware datetime, based on tz-rs and tzdb.

ToA

Serialized representation of a timestamp using a ten-element array of datetime components.

TzOutOfRangeError

Error that indicates that a provided value is outside the allowed range of values. Generally seen when constructing an offset that is greater than the allowed range.

TzStringError

Error that indicates that the provided string cannot be used in the creation of a timezone.

Enums

TimeError

A wrapper around some of the errors provided by tz-rs.

Constants

MAX_OFFSET_SECONDS

The maximum allowed offset in seconds from UTC in the future for a fixed offset.

MIN_OFFSET_SECONDS

The maximum allowed offset in seconds from UTC in the past for a fixed offset.

Type Aliases

Result

Alias for std::result::Result with the unified TimeError.