Struct spinoso_time::tzrs::Time

pub struct Time { /* private fields */ }

A wrapper around tz::datetime::DateTime which contains everything needed for date creation and conversion to match the ruby spec. Seconds and Subseconds are stored independently as i64 and u32 respectively, which gives enough granularity to meet the ruby Time spec.

Implementations

impl Time

pub fn local(
year: i32,
month: u8,
month_day: u8,
hour: u8,
minute: u8,
second: u8,
nanoseconds: u32
) -> Self

Returns a Time based on the provided values in the local timezone

Can be used to implement ruby Time#local, Time#mktime

pub fn utc(
year: i32,
month: u8,
month_day: u8,
hour: u8,
minute: u8,
second: u8,
nanoseconds: u32
) -> Self

Returns a Time based on the provided values in UTC

Can be used to implement ruby Time#utc, Time#gm

impl Time

pub fn strftime(_: Self, _format: &str) -> String

Formats time according to the directives in the given format string.

Can be used to implement Time

Panics

Panics on every invocation. Functionality is not implemented.

pub fn to_array(self) -> ToA

Serialize a Time into its components as a ToA.

ToA stores a Time as a ten-element struct of time components: [sec, min, hour, day, month, year, wday, yday, isdst, zone].

The ordering of the properties is important for the Ruby Time#to_a API.

Can be used to implement Time#to_a

Examples

let now = Time::now();
let to_array = now.to_array();
assert_eq!(to_array.sec, now.second());
assert_eq!(to_array.wday, now.day_of_week());

impl Time

pub fn round(&self, digits: u32) -> Self

Rounds sub seconds to a given precision in decimal digits (0 digits by default). It returns a new Time object. ndigits should be zero or a positive integer.

Can be used to implement Time#round

Examples

use spinoso_time::tzrs::Time;
let now = Time::local(2010, 3, 30, 5, 43, 25, 123456789);
let rounded = now.round(5);
assert_eq!(now.utc_offset(), rounded.utc_offset());
assert_eq!(123460000, rounded.nanoseconds());

impl Time

pub fn nanoseconds(&self) -> u32

Returns the number of nanoseconds for time.

The lowest digits of to_f and nsec are different because IEEE 754 double is not accurate enough to represent the exact number of nanoseconds since the Epoch.

Can be used to implement Time#nsec and Time#tv_nsec.

Examples

use spinoso_time::{tzrs::Time, NANOS_IN_SECOND};
let t = Time::utc(2022, 1, 1, 12, 0, 0, 1);
let t_float = t.to_float();
let float_nanos = (t_float - t_float.round()) * NANOS_IN_SECOND as f64;
assert_ne!(float_nanos, 1f64);
assert_eq!(t.nanoseconds(), 1);

pub fn microseconds(&self) -> u32

Returns the number of microseconds for time.

Can be used to implement Time#usec and Time#tv_usec

Examples

use spinoso_time::{tzrs::Time, MICROS_IN_NANO};
let t = Time::utc(2022, 1, 1, 12, 0, 0, 1 * MICROS_IN_NANO);
assert_eq!(t.microseconds(), 1);

pub fn second(&self) -> u8

Returns the second of the minute (0..60) for time.

Seconds range from zero to 60 to allow the system to inject leap seconds.

Can be used to implement Time#sec.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let second_of_minute = now.second();
assert_eq!(second_of_minute, 56);

pub fn minute(&self) -> u8

Returns the minute of the hour (0..59) for time.

Can be used to implement Time.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let minute_of_hour = now.minute();
assert_eq!(minute_of_hour, 34);

pub fn hour(&self) -> u8

Returns the hour of the day (0..23) for time.

Can be used to implement Time.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let hour_of_day = now.hour();
assert_eq!(hour_of_day, 12);

pub fn day(&self) -> u8

Returns the day of the month (1..n) for time.

Can be used to implement Time#day and Time#mday.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let day_of_month = now.day();
assert_eq!(day_of_month, 8);

pub fn month(&self) -> u8

Returns the month of the year (1..12) for time.

Can be used to implement Time#mon and Time#month.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let month_of_year = now.month();
assert_eq!(month_of_year, 7);

pub fn year(&self) -> i32

Returns the year for time (including the century).

Can be used to implement Time#year.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert_eq!(now.year(), 2022);

pub fn time_zone(&self) -> &str

Returns the name of the time zone as a string.

Note: UTC is an empty string due to the [UTC LocaleTimeType] being constructed with None, which is later coerced into an empty string.

Examples

use spinoso_time::tzrs::Time;
let now_utc = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert_eq!("UTC", now_utc.time_zone());

pub fn is_utc(&self) -> bool

Returns true if the time zone is UTC.

Can be used to implement Time#utc? and Time#gmt?.

use spinoso_time::tzrs::Time;
let now_utc = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert!(now_utc.is_utc());

pub fn utc_offset(&self) -> i32

Returns the offset in seconds between the timezone of time and UTC.

Can be used to implement Time#utc_offset and Time#gmt_offset.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert_eq!(now.utc_offset(), 0);

pub fn is_dst(&self) -> bool

Returns true if time occurs during Daylight Saving Time in its time zone.

Can be used to implement Time#dst? and Time#isdst.

Examples

use spinoso_time::tzrs::{Time, Offset};
use tzdb::time_zone::{europe::AMSTERDAM, pacific::AUCKLAND};
let now_ams = Time::new(2022, 5, 18, 16, 0, 0, 0, Offset::from(AMSTERDAM));
assert!(now_ams.is_dst());
let now_auckland = Time::new(2022, 5, 18, 16, 0, 0, 0, Offset::from(AUCKLAND));
assert!(!now_auckland.is_dst());

pub fn day_of_week(&self) -> u8

Returns an integer representing the day of the week, 0..6, with Sunday == 0.

Can be used to implement Time#wday.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert_eq!(now.day_of_week(), 5);

pub fn is_sunday(&self) -> bool

Returns true if time represents Sunday.

Can be used to implement Time#sunday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 4, 0, 0, 0, 0);
assert!(now.is_sunday());

pub fn is_monday(&self) -> bool

Returns true if time represents Monday.

Can be used to implement Time#monday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 5, 0, 0, 0, 0);
assert!(now.is_monday());

pub fn is_tuesday(&self) -> bool

Returns true if time represents Tuesday.

Can be used to implement Time#tuesday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 6, 0, 0, 0, 0);
assert!(now.is_tuesday());

pub fn is_wednesday(&self) -> bool

Returns true if time represents Wednesday.

Can be used to implement Time#wednesday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 7, 0, 0, 0, 0);
assert!(now.is_wednesday());

pub fn is_thursday(&self) -> bool

Returns true if time represents Thursday.

Can be used to implement Time#thursday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 1, 0, 0, 0, 0);
assert!(now.is_thursday());

pub fn is_friday(&self) -> bool

Returns true if time represents Friday.

Can be used to implement Time#friday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 2, 0, 0, 0, 0);
assert!(now.is_friday());

pub fn is_saturday(&self) -> bool

Returns true if time represents Saturday.

Can be used to implement Time#saturday?

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 3, 0, 0, 0, 0);
assert!(now.is_saturday());

pub fn day_of_year(&self) -> u16

Returns an integer representing the day of the year, 1..366.

Can be used to implement Time#yday.

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert_eq!(now.day_of_year(), 188);

impl Time

pub fn to_offset(&self, offset: Offset) -> Self

Returns a new Time object representing time based on the provided offset.

Can be used to implement Time#getlocal with a string/number parameter.

#Examples

use spinoso_time::tzrs::Time;
let local_offset = Time::now().utc_offset();
let now_utc = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let now_local = now_utc.to_local();
assert_eq!(now_local.utc_offset(), local_offset);

pub fn to_utc(&self) -> Self

Returns a new time in UTC.

Can be used to implement Time#getutc and Time#getgm.

#Examples

use spinoso_time::tzrs::Time;
let now_local = Time::now();
let now_utc = now_local.to_utc();
assert_eq!(now_utc.utc_offset(), 0);

pub fn to_local(&self) -> Self

Returns a new Time object representing time in local time (using the local time zone in effect for this process).

Can be used to implement Time#getlocal.

#Examples

use spinoso_time::tzrs::Time;
let local_offset = Time::now().utc_offset();
let now_utc = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let now_local = now_utc.to_local();
assert_eq!(now_local.utc_offset(), local_offset);

impl Time

pub fn set_offset(&mut self, offset: Offset)

Converts time to the provided time zone, modifying the receiver.

Examples

TODO

pub fn set_local(&mut self)

Converts time to local time (using the local time zone in effective at the creation time of time) modifying the receiver.

Can be used to implement Time#localtime without a parameter.

Examples

use spinoso_time::tzrs::Time;
let mut now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
let now_utc_unix = now.to_int();
now.set_local();
let now_local_unix = now.to_int();
assert_eq!(now_utc_unix, now_local_unix);

pub fn set_utc(&mut self)

Converts time to UTC (GMT), modifying the receiver.

Can be used to implement Time#utc and Time#gmtime.

Examples

use spinoso_time::tzrs::Time;
let mut now = Time::local(2022, 7, 8, 12, 34, 56, 0);
let now_local_unix = now.to_int();
now.set_local();
let now_utc_unix = now.to_int();
assert_eq!(now_local_unix, now_utc_unix);

pub fn set_offset_from_utc(&mut self, offset: Offset)

Converts time to the GMT time zone with the provided offset.

Can be used to implement Time#localtime with an offset parameter.

Examples

use spinoso_time::tzrs::{Time, Offset};
let mut now = Time::utc(2022, 7, 8, 12, 34, 56, 0);
assert!(now.is_utc());
let offset = Offset::from(3600);
now.set_offset_from_utc(offset);
assert!(!now.is_utc());
assert_eq!(now.utc_offset(), 3600);

impl Time

pub fn new(
year: i32,
month: u8,
day: u8,
hour: u8,
minute: u8,
second: u8,
nanoseconds: u32,
offset: Offset
) -> Self

Returns a new Time from the given values in the provided offset.

Can be used to implment ruby Time#new (using a Timezone Object).

Note: During DST transitions, a specific time can be ambiguous. This method will always pick the earliest date.

Examples

use spinoso_time::tzrs::{Time, Offset};
use tzdb::time_zone::pacific::AUCKLAND;
let offset = Offset::tz(AUCKLAND);
let t = Time::new(2022, 9, 25, 1, 30, 0, 0, offset);

pub fn now() -> Self

Returns a Time with the current time in the System Timezone.

Can be used to implement ruby Time#now

Examples

use spinoso_time::tzrs::Time;
let now = Time::now();

pub fn with_timespec_and_offset(
seconds: i64,
nanoseconds: u32,
offset: Offset
) -> Self

Returns a Time in the given timezone with the number of seconds and nanoseconds since the Epoch in the specified timezone.

Can be used to implement ruby Time#at

Examples

use spinoso_time::tzrs::{Time, Offset};
let offset = Offset::utc();
let t = Time::with_timespec_and_offset(0, 0, offset);
assert_eq!(t.to_int(), 0);

impl Time

pub fn to_int(&self) -> i64

Returns the number of seconds as a signed integer since the Epoch.

This function can be used to implement the ruby methods Time#to_i and Time#tv_sec

Examples

use spinoso_time::tzrs::Time;
let t = Time::utc(1970, 1, 1, 0, 1, 0, 0);
assert_eq!(t.to_int(), 60)

pub fn to_float(&self) -> f64

Returns the number of seconds since the Epoch with fractional nanos included at IEEE 754-2008 accuracy.

This function can be used to implement the ruby method Time#to_f

Examples

use spinoso_time::tzrs::Time;
let now = Time::utc(1970, 1, 1, 0, 1, 0, 1000);
assert_eq!(now.to_float(), 60.000001)

pub fn subsec_fractional(&self) -> (u32, u32)

Returns the numerator and denominator for the number of nano seconds of the Time struct unsimplified.

This can be used directly to implement Time#subsec.

This function can be used in combination with to_int to implement Time#to_r.

#Examples

use spinoso_time::tzrs::Time;
let t = Time::utc(1970, 1, 1, 0, 0, 1, 1000);
assert_eq!(t.subsec_fractional(), (1000, 1000000000));